+++ hefquin-access/src/main/java/se/liu/ida/hefquin/federation/catalog/FederationDescriptionReader.java
@@ -110,15 +110,17 @@ public class FederationDescriptionReader
 				throw new IllegalArgumentException( "Illegal serviceURI value: " + st.getObject().toString() );
 			}
 
-			membersByURI.put( serviceURI,
-			                  createFederationMember(fedMember, serviceURI, fd) );
+			final FederationMember fm = createFederationMember( fedMember,
+			                                                    NodeFactory.createURI(serviceURI),
+			                                                    fd );
+			membersByURI.put(serviceURI, fm);
 		}
 
 		return new FederationCatalogImpl(membersByURI);
 	}
 
 	protected FederationMember createFederationMember( final Resource fedMember,
-	                                                   final String serviceURI,
+	                                                   final Node serviceURI,
 	                                                   final Model fd ) {
 		final VocabularyMapping vocabMap = parseVocabMapping(fedMember, fd);
 
@@ -134,7 +136,7 @@ public class FederationDescriptionReader
 		}
 		else if ( ifaceType.equals(FDVocab.FragmentInterface) )
 		{
-			return handleFragmentInterface( iface, authInfo, protocol, vocabMap );
+			return handleFragmentInterface( iface, protocol, authInfo, vocabMap, serviceURI );
 		}
 		else if ( ifaceType.equals(FDVocab.TemplateBasedInterface) )
 		{
@@ -190,7 +192,7 @@ public class FederationDescriptionReader
 	                                                         final VocabularyMapping vocabMap,
 	                                                         final Resource fedMember,
 	                                                         final Model fd,
-	                                                         final String serviceURI ) {
+	                                                         final Node serviceURI ) {
 		if ( protocol.equals(FDVocab.SPARQLProtocol) ) {
 			final String addrStr = getSingleURIProperty(
 				iface,
@@ -198,7 +200,7 @@ public class FederationDescriptionReader
 				"SPARQL endpointAddress is required!",
 				"More than one SPARQL endpointAddress!" );
 
-			return createSPARQLEndpoint(addrStr, authInfo, vocabMap );
+			return createSPARQLEndpoint(serviceURI, addrStr, authInfo, vocabMap);
 		}
 
 		if ( protocol.equals(FDVocab.GenericWebAPIProtocol) ) {
@@ -216,7 +218,7 @@ public class FederationDescriptionReader
 			if ( trMaps.isEmpty() )
 				throw new IllegalArgumentException("The wrapped REST endpoint with service URI <" + serviceURI + "> does not have any RML triples maps.");
 
-			return createWrappedRESTEndpoint(addrStr, authInfo, null, trMaps );
+			return createWrappedRESTEndpoint(serviceURI, addrStr, null, authInfo, trMaps);
 		}
 
 		if ( protocol.equals(FDVocab.BoltProtocol) ) {
@@ -229,7 +231,7 @@ public class FederationDescriptionReader
 				"Bolt endpointAddress is required!",
 				"More than one Bolt endpointAddress!" );
 
-			return createNeo4jServer(addrStr);
+			return createNeo4jServer(serviceURI, addrStr);
 		}
 
 		if ( protocol.equals(FDVocab.GraphQLProtocol) ) {
@@ -242,7 +244,7 @@ public class FederationDescriptionReader
 				"GraphQL endpointAddress is required!",
 				"More than one GraphQL endpointAddress!" );
 
-			return createGraphQLServer(addrStr);
+			return createGraphQLServer(serviceURI, addrStr);
 		}
 
 		throw new IllegalArgumentException( protocol.toString() );
@@ -252,9 +254,10 @@ public class FederationDescriptionReader
 	 * Creates a federation member for a fragment interface (TPF, brTPF).
 	 */
 	protected FederationMember handleFragmentInterface( final Resource iface,
-	                                                    final AuthenticationInformation authInfo,
 	                                                    final Resource protocol,
-	                                                    final VocabularyMapping vocabMap ) {
+	                                                    final AuthenticationInformation authInfo,
+	                                                    final VocabularyMapping vocabMap,
+	                                                    final Node serviceURI ) {
 		if ( protocol.equals(FDVocab.TPFProtocol) ) {
 			final String addrStr = getSingleURIProperty(
 				iface,
@@ -262,7 +265,7 @@ public class FederationDescriptionReader
 				"TPF exampleFragmentAddress is required!",
 				"More than one TPF exampleFragmentAddress!" );
 
-			return createTPFServer(addrStr, authInfo, vocabMap);
+			return createTPFServer(serviceURI, addrStr, authInfo, vocabMap);
 		}
 		else if ( protocol.equals(FDVocab.brTPFProtocol) ) {
 			final String addrStr = getSingleURIProperty(
@@ -271,7 +274,7 @@ public class FederationDescriptionReader
 				"brTPF exampleFragmentAddress is required!",
 				"More than one brTPF exampleFragmentAddress!" );
 
-			return createBRTPFServer(addrStr, authInfo, vocabMap);
+			return createBRTPFServer(serviceURI, addrStr, authInfo, vocabMap);
 		}
 		else {
 			throw new IllegalArgumentException( protocol.toString() );
@@ -288,7 +291,7 @@ public class FederationDescriptionReader
 	                                                    final VocabularyMapping vocabMap,
 	                                                    final Resource fedMember,
 	                                                    final Model fd,
-	                                                    final String serviceURI ) {
+	                                                    final Node serviceURI ) {
 		if ( protocol.equals(FDVocab.GenericWebAPIProtocol) ) {
 			if ( vocabMap != null )
 				throw new IllegalArgumentException("REST APIs cannot have a vocabulary mapping.");
@@ -350,7 +353,7 @@ public class FederationDescriptionReader
 			if ( trMaps.isEmpty() )
 				throw new IllegalArgumentException("The wrapped REST endpoint with service URI <" + serviceURI + "> does not have any RML triples maps.");
 
-			return createWrappedRESTEndpoint(uriTemplateString, authInfo, params, trMaps );
+			return createWrappedRESTEndpoint(serviceURI, uriTemplateString, params, authInfo, trMaps);
 		}
 		else {
 			throw new IllegalArgumentException( protocol.toString() );
@@ -359,7 +362,7 @@ public class FederationDescriptionReader
 
 	protected List<MappingExpression> parseRMLMapping( final Resource fedMember,
 	                                                   final Model fd,
-	                                                   final String serviceURI ) {
+	                                                   final Node serviceURI ) {
 		final Resource wrapper = ModelUtils.getSingleMandatoryResourceProperty( fedMember, FDVocab.wrapper );
 
 		// TODO: Each of the TrMap-expressions created in the following
@@ -383,7 +386,7 @@ public class FederationDescriptionReader
 					                                    null );
 				}
 				catch ( final RMLParserException e ) {
-					throw new IllegalArgumentException("There is a problem in the RML mapping for <" + serviceURI + ">: " +  e.getMessage(), e );
+					throw new IllegalArgumentException("There is a problem in the RML mapping for <" + serviceURI.toString() + ">: " +  e.getMessage(), e );
 				}
 
 				trMaps.add(trMap);
@@ -439,27 +442,38 @@ public class FederationDescriptionReader
 		throw new IllegalArgumentException( "Unsupported security scheme: " + securityScheme );
 	}
 
-	protected FederationMember createSPARQLEndpoint( final String uri, final AuthenticationInformation authInfo, final VocabularyMapping vm) {
+	protected FederationMember createSPARQLEndpoint( final Node serviceURI,
+	                                                 final String uri,
+	                                                 final AuthenticationInformation authInfo,
+	                                                 final VocabularyMapping vm ) {
 		verifyExpectedURI(uri);
-		return new SPARQLEndpointImpl(uri, authInfo, vm );
+		return new SPARQLEndpointImpl( serviceURI, uri, authInfo, vm );
 	}
 
-	protected FederationMember createTPFServer( final String uri, final AuthenticationInformation authInfo, final VocabularyMapping vm ) {
+	protected FederationMember createTPFServer( final Node serviceURI,
+	                                            final String uri,
+	                                            final AuthenticationInformation authInfo,
+	                                            final VocabularyMapping vm ) {
 		verifyExpectedURI(uri);
-		return new TPFServerImpl(uri, authInfo, vm);
+		return new TPFServerImpl(serviceURI, uri, authInfo, vm);
 	}
 
-	protected FederationMember createBRTPFServer( final String uri, final AuthenticationInformation authInfo, final VocabularyMapping vm ) {
+	protected FederationMember createBRTPFServer( final Node serviceURI,
+	                                              final String uri,
+	                                              final AuthenticationInformation authInfo,
+	                                              final VocabularyMapping vm ) {
 		verifyExpectedURI(uri);
-		return new BRTPFServerImpl(uri, authInfo, vm);
+		return new BRTPFServerImpl(serviceURI, uri, authInfo, vm);
 	}
 
-	protected FederationMember createNeo4jServer( final String uri ) {
+	protected FederationMember createNeo4jServer( final Node serviceURI,
+	                                              final String uri ) {
 		verifyExpectedURI(uri);
-		return new Neo4jServerImpl(uri);
+		return new Neo4jServerImpl(serviceURI, uri);
 	}
 
-	protected FederationMember createGraphQLServer( final String uri ) {
+	protected FederationMember createGraphQLServer( final Node serviceURI,
+	                                                final String uri ) {
 		verifyExpectedURI(uri);
 
 		final GraphQLSchemaInitializer init = new GraphQLSchemaInitializerImpl();
@@ -477,18 +491,19 @@ public class FederationDescriptionReader
 			throw new IllegalArgumentException(e);
 		}
 
-		return new GraphQLEndpointImpl(uri, schema);
+		return new GraphQLEndpointImpl(serviceURI, uri, schema);
 	}
 
-	protected FederationMember createWrappedRESTEndpoint( final String uri,
-	                                                      final AuthenticationInformation authInfo,
+	protected FederationMember createWrappedRESTEndpoint( final Node serviceURI,
+	                                                      final String uri,
 	                                                      final List<RESTEndpoint.Parameter> params,
+	                                                      final AuthenticationInformation authInfo,
 	                                                      final List<MappingExpression> trMaps ) {
 		assert ! trMaps.isEmpty();
 
 		if ( trMaps.size() == 1 ) {
 			final MappingExpression expr = trMaps.get(0);
-			return new WrappedRESTEndpointImpl(uri, authInfo, params, expr );
+			return new WrappedRESTEndpointImpl(serviceURI, uri, params, authInfo, expr);
 		}
 
 		final MappingExpression[] exprs = new MappingExpression[ trMaps.size() ];
@@ -501,7 +516,7 @@ public class FederationDescriptionReader
 		final MappingExpression expr = MappingExpressionFactory.create(
 				MappingOpUnion.getInstance(),
 				exprs );
-		return new WrappedRESTEndpointImpl(uri, authInfo, params, expr );
+		return new WrappedRESTEndpointImpl(serviceURI, uri, params, authInfo, expr);
 	}
 
 	/**