+++ hefquin-access/src/main/java/se/liu/ida/hefquin/federation/members/impl/BaseForFederationMember.java
@@ -1,5 +1,7 @@
 package se.liu.ida.hefquin.federation.members.impl;
 
+import org.apache.jena.graph.Node;
+
 import se.liu.ida.hefquin.federation.FederationMember;
 
 /**
@@ -13,10 +15,22 @@ public abstract class BaseForFederationMember implements FederationMember
 	private static int counter = 0;
 
 	protected final int id;
+	protected final String serviceURI;
 
-	protected BaseForFederationMember() { id = ++counter; }
+	protected BaseForFederationMember( final Node serviceURI ) {
+		this( serviceURI.getURI() );
+	}
+
+	private BaseForFederationMember( final String serviceURI ) {
+		assert serviceURI != null;
+
+		this.serviceURI = serviceURI;
+		this.id = ++counter;
+	}
 
 	@Override public int getID() { return id; }
+
+	@Override public String getServiceURI() { return serviceURI; }
 
 	@Override
 	public boolean equals( final Object o ) {
@@ -24,6 +38,7 @@ public abstract class BaseForFederationMember implements FederationMember
 			return true;
 
 		return    o instanceof BaseForFederationMember fm
-		       && fm.getID() == id;
+		       && fm.getID() == id
+		       && fm.getServiceURI().equals(serviceURI);
 	}
 }