+++ hefquin-access/src/main/java/se/liu/ida/hefquin/federation/members/impl/RESTEndpointImpl.java
@@ -2,6 +2,8 @@ package se.liu.ida.hefquin.federation.members.impl;
 
 import java.util.List;
 
+import org.apache.jena.graph.Node;
+
 import se.liu.ida.hefquin.base.query.SPARQLGraphPattern;
 import se.liu.ida.hefquin.federation.authentication.AuthenticationInformation;
 import se.liu.ida.hefquin.federation.members.RESTEndpoint;
@@ -13,7 +15,12 @@ public class RESTEndpointImpl extends BaseForFederationMember
 	protected final List<RESTEndpoint.Parameter> params;
 	protected final AuthenticationInformation authInfo;
 
-	public RESTEndpointImpl( final String urlTemplate, final AuthenticationInformation authInfo, final List<RESTEndpoint.Parameter> params ) {
+	public RESTEndpointImpl( final Node serviceURI,
+	                         final String urlTemplate,
+	                         final List<RESTEndpoint.Parameter> params,
+	                         final AuthenticationInformation authInfo ) {
+		super(serviceURI);
+
 		assert urlTemplate != null && ! urlTemplate.isEmpty();
 		this.urlTemplate = urlTemplate;
 		this.authInfo = authInfo;