+++ hefquin-access/src/main/java/se/liu/ida/hefquin/federation/members/impl/SPARQLEndpointImpl.java
@@ -1,5 +1,7 @@
 package se.liu.ida.hefquin.federation.members.impl;
 
+import org.apache.jena.graph.Node;
+
 import se.liu.ida.hefquin.base.data.VocabularyMapping;
 import se.liu.ida.hefquin.federation.authentication.AuthenticationInformation;
 import se.liu.ida.hefquin.federation.members.SPARQLEndpoint;
@@ -10,8 +12,11 @@ public class SPARQLEndpointImpl extends BaseForRDFBasedFederationMember
 	protected final String url;
 	protected final AuthenticationInformation authInfo;
 
-	public SPARQLEndpointImpl( final String url, final AuthenticationInformation authInfo, final VocabularyMapping vm ) {
-		super(vm);
+	public SPARQLEndpointImpl( final Node serviceURI,
+	                           final String url,
+	                           final AuthenticationInformation authInfo,
+	                           final VocabularyMapping vm ) {
+		super(serviceURI, vm);
 
 		assert url != null && ! url.isEmpty();
 		this.url = url;