+++ hefquin-engine/src/main/java/se/liu/ida/hefquin/engine/queryplan/utils/TextBasedLogicalPlanPrinterImpl.java
@@ -5,6 +5,7 @@ import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
 
 import org.apache.jena.sparql.core.Var;
 import org.apache.jena.sparql.core.VarExprList;
@@ -149,6 +150,30 @@ public class TextBasedLogicalPlanPrinterImpl extends BaseForTextBasedPlanPrinter
 		@Override
 		public void visit( final LogicalOpRequest<?,?> op ) {
 			record( op.getFederationMember() );
+			record( op.getRequest() );
+
+			props.add( "may reduce duplicates: " + op.mayReduce() );
+		}
+
+		@Override
+		public void visit( final LogicalOpMultiRequest op ) {
+			props.add( "service variable: " + op.getServiceVariable().toString() );
+
+			final int n = op.getFederationMembers().size();
+			props.add( "number of fed.members: " + n );
+
+			if ( n > 0 ) {
+				final Iterator<FederationMember> it = op.getFederationMembers().iterator();
+
+				// print first member at the same line
+				props.add( "fed.members: " + it.next().toString() );
+
+				// print each of the remaining members at a separate line
+				while ( it.hasNext() ) {
+					props.add( "             " + it.next().toString() );
+				}
+			}
+
 			record( op.getRequest() );
 
 			props.add( "may reduce duplicates: " + op.mayReduce() );
@@ -303,6 +328,26 @@ public class TextBasedLogicalPlanPrinterImpl extends BaseForTextBasedPlanPrinter
 		protected void record( final DataRetrievalRequest req ) {
 			if ( req instanceof SPARQLRequest sreq ) {
 				record( sreq.getQueryPattern() );
+
+				final String prVarsString;
+				final Set<Var> prVars = sreq.getProjectionVars();
+				if ( prVars == null ) {
+					prVarsString = "*";
+				}
+				else if ( prVars.isEmpty() ) {
+					prVarsString = "-none-";
+				}
+				else {
+					final StringBuilder sb = new StringBuilder();
+					final Iterator<Var> it = prVars.iterator();
+					sb.append( it.next().toString() );
+					while ( it.hasNext() ) {
+						sb.append( ", " + it.next().toString() );
+					}
+					prVarsString = sb.toString();
+				}
+
+				props.add( "projection variables: " + prVarsString );
 			}
 			else {
 				props.add( "request (" + req.hashCode() +  "): " + req.toString() );